Add instrumented unti test for implementation of Parcelable
